+++ src/main/java/mdlaf/themes/JMarsDarkTheme.java
@@ -127,6 +127,10 @@ public class JMarsDarkTheme extends AbstractMaterialTheme {
         this.selectedIconToggleButton = MaterialImageFactory.getInstance().getImage(MaterialImageFactory.TOGGLE_BUTTON_ON_WHITE);
 
         super.iconCloseTitlePane = MaterialImageFactory.getInstance().getImage(MaterialImageFactory.CLOSE_WINDOWS_WHITE);
+
+        super.leafIconTree = MaterialImageFactory.getInstance().getImage(
+                MaterialImageFactory.BLANK
+        );
     }
 
     @Override